+++ src/test/java/org/scify/jedai/configuration/StepByStepGridConfigurationCCER.java
@@ -48,8 +48,8 @@ import org.scify.jedai.utilities.datastructures.BilateralDuplicatePropagation;
 public class StepByStepGridConfigurationCCER {
 
     static float getTotalComparisons(List<AbstractBlock> blocks) {
-        float originalComparisons = 0;
-        originalComparisons = blocks.stream().map(AbstractBlock::getNoOfComparisons).reduce(originalComparisons, Float::sum);
+        int originalComparisons = 0;
+        originalComparisons = blocks.stream().map(AbstractBlock::getNoOfComparisons).reduce(originalComparisons, Integer::sum);
         System.out.println("Original comparisons\t:\t" + originalComparisons);
         return originalComparisons;
     }
